+++ library/src/main/java/cloud/pace/sdk/appkit/communication/MessageHandler.kt
@@ -2,7 +2,8 @@ package cloud.pace.sdk.appkit.communication
 
 enum class MessageHandler(val id: String, val timeoutMillis: Long) {
 
-    INVALID_TOKEN("pace_invalidToken", 60_000),
+    GET_ACCESS_TOKEN("pace_getAccessToken", 60_000),
+    LOGOUT("pace_logout", 30_000),
     IMAGE_DATA("pace_imageData", 5_000),
     VERIFY_LOCATION("pace_verifyLocation", 60_000),
     BACK("pace_back", 5_000),