+++ library/src/main/java/cloud/pace/sdk/appkit/model/InvalidTokenReason.kt
@@ -1,6 +1,0 @@
-package cloud.pace.sdk.appkit.model
-
-enum class InvalidTokenReason(val value: String) {
-    UNAUTHORIZED("unauthorized"),
-    OTHER("other")
-}